requirements.txt
pynput==1.8.1
requests==2.32.3
pillow==10.4.0
google-generativeai==0.8.5
ttkbootstrap==1.13.11
python-dotenv
pywin32
openai
anthropic

